Rename contents -> table-of-contents
