+++ packages/11ty/_includes/components/table-of-contents/image.js
